Move "get accessible tabs" action to employee controller
